# apps/releases/var/www/api/app/doc/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "500762 HDMI Digital Signage Media Player Api Documentation",
        "version": "1.0.0"
    },
    "host": "device_ip:8888",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/rs232/sendData": {
            "put": {
                "tags": [
                    "Rs232"
                ],
                "summary": "rs232 send hexadecimal data",
                "description": "send hexadecimal data.",
                "operationId": "rs232_sendData",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token retrieved from the login API",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "required": [
                                "data",
                                "isFeedback"
                            ],
                            "properties": {
                                "data": {
                                    "type": "string",
                                    "example": "A055B1FF"
                                },
                                "isFeedback": {
                                    "description": "1 or 0",
                                    "type": "string",
                                    "example": "0"
                                },
                                "waitTime": {
                                    "description": "in seconds",
                                    "type": "string",
                                    "example": "1000"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "properties": {
                                        "feedbackdata": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "authentication failed",
                        "schema": {
                            "$ref": "#/definitions/401_response"
                        }
                    }
                }
            }
        },
        "/rs232/configuration": {
            "put": {
                "tags": [
                    "Rs232"
                ],
                "summary": "Set rs232 configuration",
                "description": "Set rs232 configuration",
                "operationId": "rs232_setConfiguration",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token retrieved from the login API",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "RS232Baudrate": {
                                    "description": "9600, 19200, 38400, 57600, 115200",
                                    "type": "integer",
                                    "example": "9600"
                                },
                                "RS232Databits": {
                                    "description": "7,8",
                                    "type": "integer",
                                    "example": "8"
                                },
                                "RS232Parity": {
                                    "description": "none, odd, even",
                                    "type": "string",
                                    "example": "none"
                                },
                                "RS232Stopbits": {
                                    "description": "1, 2",
                                    "type": "integer",
                                    "example": "1"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "authentication failed",
                        "schema": {
                            "$ref": "#/definitions/401_response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "401_response": {
            "properties": {
                "status": {
                    "type": "string",
                    "example": "failed"
                },
                "data": {
                    "type": "array",
                    "items": {}
                }
            }
        },
...
    },
...
}